Page 1 Functionality Completed - Page 2 Partial Completion
Updated the Flights page so it displays the flight information in the reserve input sections.

Added Reservation and Reservation Manager classes.

Added the ability to create a reservation on flights page.

Reservation code is generated when res. is created.

Add the same functionality as the flights page except for the updated of a reservation and persist method.